fix(core)!: widen Epoch to u128 for lossless leader-epoch encoding (#221)
The paxos driver packed an OmniPaxos Ballot — identity (config_id: u32,
n: u32, pid: u64), 128 bits — into a 64-bit Epoch by masking config_id and
pid to 16 bits each. Distinct ballots could collide onto the same Epoch
(e.g. pid 1 and pid 65537), and that value is load-bearing: the fence in
PaxosDriver::persist_high_water and Allocator::try_commit_window_extension
distinguishes leaders by exact equality, and the client's leader cache
orders leader observations by it. A collision lets a superseded leader's
persisted high-water pass the fence and silently advance the durable bound.

A u64 cannot injectively hold the full ballot domain, so widen Epoch to
u128. The paxos encoding becomes lossless and total — config_id<<96 |
n<<64 | pid, no truncation, no validation — while staying monotonic, which
the equality fence and the client's monotone-forward cache both rely on.
decode_epoch is now an exact inverse. priority remains intentionally
unencoded (a static per-node tiebreaker determined by pid).

The on-wire epoch carries the full value as two uint64 halves rather than
truncating at the boundary: GetTsResponse.epoch becomes epoch_hi/epoch_lo
and LeaderHint.leader_epoch becomes leader_epoch_hi/leader_epoch_lo, with
lexicographic (hi, lo) order equal to numeric order. The file and openraft
drivers zero-extend their existing u64 epochs (no behaviour change).

Property tests cover the lossless round-trip and order-preservation
(injectivity + monotonicity) over the full ballot domain, and the
to_wire/from_wire round-trip and ordering across the 64-bit boundary.

BREAKING CHANGE: the GetTs gRPC response and the leader-hint trailer split
their u64 epoch field into two uint64 halves. Clients and servers must be
upgraded together.

Closes #206

+++ crates/tsoracle-core/tests/monotonicity.rs
@@ -80,7 +80,7 @@ proptest! {
                     };
                     // committed_ceiling must be >= fence_floor for immediate serving.
                     let committed_ceiling = fence_floor.saturating_add(30_000);
-                    allocator.try_on_leadership_gained(fence_floor, committed_ceiling, Epoch(epoch_counter)).unwrap();
+                    allocator.try_on_leadership_gained(fence_floor, committed_ceiling, Epoch(u128::from(epoch_counter))).unwrap();
                 }
                 Op::Lose => {
                     allocator.on_leadership_lost();
